Keep track of number of remaining checks
Another off-board square is now allocated forkeeping track of the number
of remaining checks each side can suffer before losing. When this double
counter is 0, counting is disabled. Otherwise each check will decrease
the count of the checked side, and when it reaches 0 the check will be
treated as a checkmate. In variant 3check the counters will be initialized
to 3, otherwise to 0.