Correct promotion zone for blacked-out ranks
When the window height has to be larger than the board height, to accomodate
the holdings (as in Wa Shogi), this is done by defining the board height as
large as the holdings size, and blacking out the extra ranks. But when the
depth is calculated as 1/3 of the board height, it should use the real
board height, not the artificially enlarged one.