Implement Makruk counting rules
Adjudication is now performed based on the Makruk counting rules, assuming
the players count in the way most favorable for them. This means that when
the maximum count is N, the game is adjudicated draw immediately after the
Nth move since the counting could have started when the opponent is bared,
but only on the next move when not, to give him the opportunity to checkmate
when he was not counting. This adjudication is switched off if -ruleMoves
is set to a negative value, (normally this would be 0 in Makruk), and is
also dependent on legality testing being on.