Implement (unconditional) protocol auto-detection
By sending "uci" to the engine after the feature timeout expires,
and redoing the engine replacement with UCI ticked when "uciok" is
received, UCI engines will install properly even when they are first
assumed to be CECP engines. An extra timeout period is taken to wait
for the possible "uciok" response. If this expires too, the engine is
assumed to be WB v1.
  Adding the engine to the engine list had to be deferred until the
protocol is known. This means the "-protocolVersion 1" option can now
also automatically added to the engine line for v1 engines.