Locate engine list in separate file
The option -firstChessProgramNames, which contains the list of registered
engines, can now be kept in a dedicated settings file. This file will
be read at startup after the normal settings file and command-line
arguments. It will also be read and saved when dialogs altering the
engine list are used (Load Engine and Edit Engine List). This will
effectively cause changes in the engine list made by any XBoard instance
to be instantly broadcasted to all other instances. The name of this
engine-list file can be specified by the new persistent option -engineList.
If this is set to an empty string (the default), this new feature will
not be used, and the list in the normal settings file will be used
as always. This list is currently still written in the user settings file,
together with the other settings.