Also update engine list when saving engine settings
The 'Make Persistent' button in the Engine Settings dialog will now
also cause the -engineList file to be read before being processed,
and be written back after it, to keep the list up to date for all
open XBoard instances. There still is a problem here, that when another
instance has changed settings of the currently loaded engine, this engine
is no longerrecognized as being in the list. But at least the user will
be warned in such a case that makig the settings change persistent has
failed.