Redesign WinBoard Load Engine dialog
The checkboxes for protocol choice are replaced by a combobox, which
also contains an autodetect option. This is encoded through the kludge
of setting isUCI = 3.
 The fields that do not usually have to be specified are now put in
group boxes, which make it more clear the user can ignore them.
For this a ew control type GroupBox had to be defined, and supported
in the GenericPopup.