Add option -defaultInstallProtocol
A persistent option is added to determine the initial setting of the
protocol-selector combobox in the Load Engine dialog. It must be given
a value 0-4 corresponding to the choice order. This is mainly intended
for configuring Shogi or Xiangqi packages, to select USI or USI (for which
auto-detection does not work).