Fix RxK castling from engine in variant seirawan
The RxK notation for castling with gating at the Rook square was mistaken
for a non-compliant castling notation in games without Fischer castling,
and corrected to the compliant notation by adjusting the to-square, assumed
to be the Rook. But this backfired for RxK, where it is the King. Now
castlings with the to-square on the central file are exempted from this
correction.