Fix Edit Position with engine-defined variant
Submitting a new position starts a new game, but the position would be
set up before the engine had the time to reply to the 'variant' command
wit a 'setup'. As a consequence WinBoard would be unaware of the correct
starting position, with as a consequence that Pawns might not be considered
virgin for their Betza 'i' moves.
 Now even late reception of a 'setup' command from the engine when
startedFromSetupPosition is true will cause initialPosition to be
initialized. The assumption is that XBoard was already switched to the
variant before, so that the 'setup' command is not needed for defining the
pieceToCharTable used to interpret the FEN that set up the position.